+++ tortoise/backends/base_postgres/client.py
@@ -1,6 +1,16 @@
 import abc
 from functools import wraps
-from typing import Any, Callable, List, Optional, SupportsInt, Tuple, Type, TypeVar, Union
+from typing import (
+    Any,
+    Callable,
+    List,
+    Optional,
+    SupportsInt,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
 
 from pypika import PostgreSQLQuery